+++ src/crawlee/sessions/_session.py
@@ -199,21 +199,16 @@ class Session:
         self,
         *,
         status_code: int,
-        additional_blocked_status_codes: set[int] | None = None,
         ignore_http_error_status_codes: set[int] | None = None,
     ) -> bool:
         """Evaluate whether a session should be retired based on the received HTTP status code.
 
         Args:
             status_code: The HTTP status code received from a server response.
-            additional_blocked_status_codes: Optional additional status codes that should trigger session retirement.
             ignore_http_error_status_codes: Optional status codes to allow suppression of
             codes from `blocked_status_codes`.
 
         Returns:
             True if the session should be retired, False otherwise.
         """
-        if additional_blocked_status_codes and status_code in additional_blocked_status_codes:
-            return True
-
         return status_code in (self._blocked_status_codes - (ignore_http_error_status_codes or set()))