+++ tests/unit/http_clients/test_curl_impersonate.py
@@ -77,33 +77,34 @@ async def test_send_request_with_proxy_disabled(
 
 
 async def test_crawl_allow_redirects_by_default(http_client: CurlImpersonateHttpClient, httpbin: URL) -> None:
-    final_url = str(httpbin / 'get')
-    redirect_url = str((httpbin / 'redirect-to').with_query(url=final_url))
+    target_url = str(httpbin.with_path('get', keep_query=False))
+    check_url = str(httpbin / 'get')
+    redirect_url = str((httpbin / 'redirect-to').update_query(url=target_url))
     request = Request.from_url(redirect_url)
-
     crawling_result = await http_client.crawl(request)
 
     assert crawling_result.http_response.status_code == 200
-    assert request.loaded_url == final_url
+    assert request.loaded_url == check_url
 
 
 async def test_crawl_allow_redirects_false(httpbin: URL) -> None:
     http_client = CurlImpersonateHttpClient(allow_redirects=False)
 
-    final_url = str(httpbin / 'get')
-    redirect_url = str((httpbin / 'redirect-to').with_query(url=final_url))
+    target_url = str(httpbin.with_path('get', keep_query=False))
+    check_url = str(httpbin / 'get')
+    redirect_url = str((httpbin / 'redirect-to').update_query(url=target_url))
     request = Request.from_url(redirect_url)
 
     crawling_result = await http_client.crawl(request)
 
     assert crawling_result.http_response.status_code == 302
-    assert crawling_result.http_response.headers['Location'] == final_url
+    assert crawling_result.http_response.headers['Location'] == check_url
     assert request.loaded_url == redirect_url
 
 
 async def test_send_request_allow_redirects_by_default(http_client: CurlImpersonateHttpClient, httpbin: URL) -> None:
-    final_url = str(httpbin / 'get')
-    redirect_url = str((httpbin / 'redirect-to').with_query(url=final_url))
+    target_url = str(httpbin.with_path('get', keep_query=False))
+    redirect_url = str((httpbin / 'redirect-to').update_query(url=target_url))
 
     response = await http_client.send_request(redirect_url)
 
@@ -113,10 +114,11 @@ async def test_send_request_allow_redirects_by_default(http_client: CurlImperson
 async def test_send_request_allow_redirects_false(httpbin: URL) -> None:
     http_client = CurlImpersonateHttpClient(allow_redirects=False)
 
-    final_url = str(httpbin / 'get')
-    redirect_url = str((httpbin / 'redirect-to').with_query(url=final_url))
+    target_url = str(httpbin.with_path('get', keep_query=False))
+    check_url = str(httpbin / 'get')
+    redirect_url = str((httpbin / 'redirect-to').update_query(url=target_url))
 
     response = await http_client.send_request(redirect_url)
 
     assert response.status_code == 302
-    assert response.headers['Location'] == final_url
+    assert response.headers['Location'] == check_url